requirements.txt
JPype1
konlpy
pandas
emosent-py